+++ Arcane Waters - Alpha/Assets/Project Tools/PlayerClassDataTool/StatHolderPanel.cs
@@ -14,82 +14,82 @@ public class StatHolderPanel : MonoBehaviour {
       PlayerStats statData = new PlayerStats();
 
       // Per Level stats
-      statData.userDefaultStats.hpPerLevel = int.Parse(_hpPerLevel.text);
-      statData.userDefaultStats.armorPerLevel = int.Parse(_armorPerLevel.text);
-      statData.userDefaultStats.apPerLevel = int.Parse(_apPerLevel.text);
+      statData.userDefaultStats.hpPerLevel = float.Parse(_hpPerLevel.text);
+      statData.userDefaultStats.armorPerLevel = float.Parse(_armorPerLevel.text);
+      statData.userDefaultStats.apPerLevel = float.Parse(_apPerLevel.text);
 
       // Base Stats
-      statData.userDefaultStats.bonusAP = int.Parse(_addedAP.text);
-      statData.userDefaultStats.bonusMaxHP = int.Parse(_addedMaxHP.text);
-      statData.userDefaultStats.bonusArmor = int.Parse(_addedArmor.text);
-      statData.userDefaultStats.bonusATK = int.Parse(_addedATK.text);
-      statData.userDefaultStats.bonusATKPerLevel = int.Parse(_atkPerlevel.text);
+      statData.userDefaultStats.bonusAP = float.Parse(_addedAP.text);
+      statData.userDefaultStats.bonusMaxHP = float.Parse(_addedMaxHP.text);
+      statData.userDefaultStats.bonusArmor = float.Parse(_addedArmor.text);
+      statData.userDefaultStats.bonusATK = float.Parse(_addedATK.text);
+      statData.userDefaultStats.bonusATKPerLevel = float.Parse(_atkPerlevel.text);
 
       // Special Stats
-      statData.userDefaultStats.bonusINT = int.Parse(_bonusINT.text);
-      statData.userDefaultStats.bonusVIT = int.Parse(_bonusVIT.text);
-      statData.userDefaultStats.bonusPRE = int.Parse(_bonusPRE.text);
-      statData.userDefaultStats.bonusSPT = int.Parse(_bonusSPT.text);
-      statData.userDefaultStats.bonusLUK = int.Parse(_bonusLUK.text);
-      statData.userDefaultStats.bonusSTR = int.Parse(_bonusSTR.text);
+      statData.userDefaultStats.bonusINT = float.Parse(_bonusINT.text);
+      statData.userDefaultStats.bonusVIT = float.Parse(_bonusVIT.text);
+      statData.userDefaultStats.bonusPRE = float.Parse(_bonusPRE.text);
+      statData.userDefaultStats.bonusSPT = float.Parse(_bonusSPT.text);
+      statData.userDefaultStats.bonusLUK = float.Parse(_bonusLUK.text);
+      statData.userDefaultStats.bonusSTR = float.Parse(_bonusSTR.text);
 
       // Special Stats per level
-      statData.userDefaultStats.bonusINTperLevel = int.Parse(_bonusINTperLevel.text);
-      statData.userDefaultStats.bonusVITperLevel = int.Parse(_bonusVITperLevel.text);
-      statData.userDefaultStats.bonusPREperLevel = int.Parse(_bonusPREperLevel.text);
-      statData.userDefaultStats.bonusSPTperLevel = int.Parse(_bonusSPTperLevel.text);
-      statData.userDefaultStats.bonusLUKperLevel = int.Parse(_bonusLUKperLevel.text);
-      statData.userDefaultStats.bonusSTRperLevel = int.Parse(_bonusSTRperLevel.text);
+      statData.userDefaultStats.bonusINTperLevel = float.Parse(_bonusINTperLevel.text);
+      statData.userDefaultStats.bonusVITperLevel = float.Parse(_bonusVITperLevel.text);
+      statData.userDefaultStats.bonusPREperLevel = float.Parse(_bonusPREperLevel.text);
+      statData.userDefaultStats.bonusSPTperLevel = float.Parse(_bonusSPTperLevel.text);
+      statData.userDefaultStats.bonusLUKperLevel = float.Parse(_bonusLUKperLevel.text);
+      statData.userDefaultStats.bonusSTRperLevel = float.Parse(_bonusSTRperLevel.text);
 
       // Defenses Stats
-      statData.userCombatStats.bonusResistancePhys = int.Parse(_bonusDEFPhys.text);
-      statData.userCombatStats.bonusResistanceFire = int.Parse(_bonusDEFFire.text);
-      statData.userCombatStats.bonusResistanceEarth = int.Parse(_bonusDEFEarth.text);
-      statData.userCombatStats.bonusResistanceWind = int.Parse(_bonusDEFWind.text);
-      statData.userCombatStats.bonusResistanceWater = int.Parse(_bonusDEFWater.text);
-      statData.userCombatStats.bonusResistanceAll = int.Parse(_bonusDEFAll.text);
+      statData.userCombatStats.bonusResistancePhys = float.Parse(_bonusDEFPhys.text);
+      statData.userCombatStats.bonusResistanceFire = float.Parse(_bonusDEFFire.text);
+      statData.userCombatStats.bonusResistanceEarth = float.Parse(_bonusDEFEarth.text);
+      statData.userCombatStats.bonusResistanceWind = float.Parse(_bonusDEFWind.text);
+      statData.userCombatStats.bonusResistanceWater = float.Parse(_bonusDEFWater.text);
+      statData.userCombatStats.bonusResistanceAll = float.Parse(_bonusDEFAll.text);
 
       // Defenses Stats per level
-      statData.userCombatStats.bonusResistancePhysPerLevel = int.Parse(_bonusDEFPhysPerLevel.text);
-      statData.userCombatStats.bonusResistanceFirePerLevel = int.Parse(_bonusDEFFirePerLevel.text);
-      statData.userCombatStats.bonusResistanceEarthPerLevel = int.Parse(_bonusDEFEarthPerLevel.text);
-      statData.userCombatStats.bonusResistanceWindPerLevel = int.Parse(_bonusDEFWindPerLevel.text);
-      statData.userCombatStats.bonusResistanceWaterPerLevel = int.Parse(_bonusDEFWaterPerLevel.text);
-      statData.userCombatStats.bonusResistanceAllPerLevel = int.Parse(_bonusDEFAllPerLevel.text);
+      statData.userCombatStats.bonusResistancePhysPerLevel = float.Parse(_bonusDEFPhysPerLevel.text);
+      statData.userCombatStats.bonusResistanceFirePerLevel = float.Parse(_bonusDEFFirePerLevel.text);
+      statData.userCombatStats.bonusResistanceEarthPerLevel = float.Parse(_bonusDEFEarthPerLevel.text);
+      statData.userCombatStats.bonusResistanceWindPerLevel = float.Parse(_bonusDEFWindPerLevel.text);
+      statData.userCombatStats.bonusResistanceWaterPerLevel = float.Parse(_bonusDEFWaterPerLevel.text);
+      statData.userCombatStats.bonusResistanceAllPerLevel = float.Parse(_bonusDEFAllPerLevel.text);
 
       // Damage Stats
-      statData.userCombatStats.bonusDamagePhys = int.Parse(_bonusATKPhys.text);
-      statData.userCombatStats.bonusDamageFire = int.Parse(_bonusATKFire.text);
-      statData.userCombatStats.bonusDamageEarth = int.Parse(_bonusATKEarth.text);
-      statData.userCombatStats.bonusDamageWind = int.Parse(_bonusATKWind.text);
-      statData.userCombatStats.bonusDamageWater = int.Parse(_bonusATKWater.text);
-      statData.userCombatStats.bonusDamageAll = int.Parse(_bonusATKAll.text);
+      statData.userCombatStats.bonusDamagePhys = float.Parse(_bonusATKPhys.text);
+      statData.userCombatStats.bonusDamageFire = float.Parse(_bonusATKFire.text);
+      statData.userCombatStats.bonusDamageEarth = float.Parse(_bonusATKEarth.text);
+      statData.userCombatStats.bonusDamageWind = float.Parse(_bonusATKWind.text);
+      statData.userCombatStats.bonusDamageWater = float.Parse(_bonusATKWater.text);
+      statData.userCombatStats.bonusDamageAll = float.Parse(_bonusATKAll.text);
 
       // Damage Stats per level
-      statData.userCombatStats.bonusDamagePhysPerLevel = int.Parse(_bonusATKPhysPerLevel.text);
-      statData.userCombatStats.bonusDamageFirePerLevel = int.Parse(_bonusATKFirePerLevel.text);
-      statData.userCombatStats.bonusDamageEarthPerLevel = int.Parse(_bonusATKEarthPerLevel.text);
-      statData.userCombatStats.bonusDamageWindPerLevel = int.Parse(_bonusATKWindPerLevel.text);
-      statData.userCombatStats.bonusDamageWaterPerLevel = int.Parse(_bonusATKWaterPerLevel.text);
-      statData.userCombatStats.bonusDamageAllPerLevel = int.Parse(_bonusATKAllPerLevel.text);
+      statData.userCombatStats.bonusDamagePhysPerLevel = float.Parse(_bonusATKPhysPerLevel.text);
+      statData.userCombatStats.bonusDamageFirePerLevel = float.Parse(_bonusATKFirePerLevel.text);
+      statData.userCombatStats.bonusDamageEarthPerLevel = float.Parse(_bonusATKEarthPerLevel.text);
+      statData.userCombatStats.bonusDamageWindPerLevel = float.Parse(_bonusATKWindPerLevel.text);
+      statData.userCombatStats.bonusDamageWaterPerLevel = float.Parse(_bonusATKWaterPerLevel.text);
+      statData.userCombatStats.bonusDamageAllPerLevel = float.Parse(_bonusATKAllPerLevel.text);
 
       // Ship Data
-      statData.userShipStats.bonusShipDamage = int.Parse(_damageBase.text);
-      statData.userShipStats.bonusShipDamagePerLevel = int.Parse(_damagePerLevel.text);
-      statData.userShipStats.bonusShipHealth = int.Parse(_healthBase.text);
-      statData.userShipStats.bonusShipHealthPerLevel = int.Parse(_healthPerLevel.text);
+      statData.userShipStats.bonusShipDamage = float.Parse(_damageBase.text);
+      statData.userShipStats.bonusShipDamagePerLevel = float.Parse(_damagePerLevel.text);
+      statData.userShipStats.bonusShipHealth = float.Parse(_healthBase.text);
+      statData.userShipStats.bonusShipHealthPerLevel = float.Parse(_healthPerLevel.text);
 
-      statData.userShipStats.bonusShipRange = int.Parse(_rangeBase.text);
-      statData.userShipStats.bonusShipRangePerLevel = int.Parse(_rangePerLevel.text);
-      statData.userShipStats.bonusShipSpeed = int.Parse(_speedBase.text);
-      statData.userShipStats.bonusShipSpeedPerLevel = int.Parse(_speedPerLevel.text);
+      statData.userShipStats.bonusShipRange = float.Parse(_rangeBase.text);
+      statData.userShipStats.bonusShipRangePerLevel = float.Parse(_rangePerLevel.text);
+      statData.userShipStats.bonusShipSpeed = float.Parse(_speedBase.text);
+      statData.userShipStats.bonusShipSpeedPerLevel = float.Parse(_speedPerLevel.text);
 
-      statData.userShipStats.bonusShipSailors = int.Parse(_sailorBase.text);
-      statData.userShipStats.bonusShipSailorsPerLevel = int.Parse(_sailorPerLevel.text);
-      statData.userShipStats.bonusShipCargoRoom = int.Parse(_cargoRoomBase.text);
-      statData.userShipStats.bonusShipCargoRoomPerLevel = int.Parse(_cargoRoomPerLevel.text);
-      statData.userShipStats.bonusShipSupplyRoom = int.Parse(_supplyRoomBase.text);
-      statData.userShipStats.bonusShipSupplyRoomPerLevel = int.Parse(_supplyRoomPerLevel.text);
+      statData.userShipStats.bonusShipSailors = float.Parse(_sailorBase.text);
+      statData.userShipStats.bonusShipSailorsPerLevel = float.Parse(_sailorPerLevel.text);
+      statData.userShipStats.bonusShipCargoRoom = float.Parse(_cargoRoomBase.text);
+      statData.userShipStats.bonusShipCargoRoomPerLevel = float.Parse(_cargoRoomPerLevel.text);
+      statData.userShipStats.bonusShipSupplyRoom = float.Parse(_supplyRoomBase.text);
+      statData.userShipStats.bonusShipSupplyRoomPerLevel = float.Parse(_supplyRoomPerLevel.text);
 
       return statData;
    }
